refactor: update cycle import
