+++ docs/data/base/components/select/UseSelect.js
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import useSelect from '@mui/base/useSelect';
-
 import { styled } from '@mui/system';
 
 const grey = {
